requirements.txt
pipenv
requests
numpy
pandas
tabula-py
tqdm
xlsxwriter
tablib
flask
bizdays